move toolbox to wrapper component
